+++ Wimc.Tests.Unit/Wimc.Tests.Unit.csproj
@@ -1,0 +1,34 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net5.0</TargetFramework>
+
+        <IsPackable>false</IsPackable>
+    </PropertyGroup>
+
+    <ItemGroup>
+        <PackageReference Include="FluentAssertions" Version="5.10.3" />
+        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.7.1" />
+        <PackageReference Include="Moq" Version="4.16.1" />
+        <PackageReference Include="xunit" Version="2.4.1" />
+        <PackageReference Include="xunit.runner.visualstudio" Version="2.4.3">
+            <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+            <PrivateAssets>all</PrivateAssets>
+        </PackageReference>
+        <PackageReference Include="coverlet.collector" Version="1.3.0">
+            <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+            <PrivateAssets>all</PrivateAssets>
+        </PackageReference>
+    </ItemGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\Mx.EntityFramework.Contracts\Mx.EntityFramework.Contracts.csproj" />
+      <ProjectReference Include="..\Mx.EntityFramework\Mx.EntityFramework.csproj" />
+      <ProjectReference Include="..\Mx.Library\Mx.Library.csproj" />
+      <ProjectReference Include="..\Wimc.Data\Wimc.Data.csproj" />
+      <ProjectReference Include="..\Wimc.Infrastructure\Wimc.Infrastructure.csproj" />
+      <ProjectReference Include="..\Wimc.Worker\Wimc.Worker.csproj" />
+      <ProjectReference Include="..\Wimc\Wimc.csproj" />
+    </ItemGroup>
+
+</Project>
