+++ Wimc.Worker/Program.cs
@@ -1,0 +1,29 @@
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
+
+namespace WImc.Worker
+{
+    public class Program
+    {
+        
+        public static void Main(string[] args)
+        {
+            CreateHostBuilder(args).Build().Run();
+        }
+
+        public static IHostBuilder CreateHostBuilder(string[] args) =>
+            Host.CreateDefaultBuilder(args)
+                .ConfigureAppConfiguration(configurationBuilder =>
+                {
+                    configurationBuilder.AddJsonFile("appsettings.json", false, reloadOnChange: true);
+                    configurationBuilder.AddJsonFile("appsettings.secrets.json", true, reloadOnChange: true);
+                })
+                .ConfigureServices((hostContext, services) =>
+                {
+                    services.AddOptions().AddWorkerDependencies(hostContext.Configuration);
+                    services.AddHostedService<Worker>();
+                    
+                });
+    }
+}
