+++ Wimc.Worker/Properties/launchSettings.json
