+++ Wimc.Worker/Worker.cs
