+++ Wimc.Worker/WorkerStartup.cs
