+++ Wimc.Worker/appsettings.json
